debug: put a delve on it
Add Makefile delve-container target that builds an image using
DelveDockerfile. This Dockerfile adds an additional step to build delve
and place it in the container, and changes the controller build to
include debug symbols and disable optimizations.

The controller starts automatically (it is run with --continue). If
testing code that runs at startup only, spawn the container, connnect,
add breakpoints, restart, and continue.